Class changes done
Added a few features of the db and the gui. Switching to new branch  to continue that development
