+++ .gitignore
@@ -25,6 +25,9 @@
 # Ignore pycache
 **/__pycache__/
 
+# Ignore Excel files (optimization results)
+*.xlsx
+
 # Ignore system files
 .DS_Store
 Thumbs.db
