+++ pso/database/data.py
@@ -4,7 +4,7 @@ import numpy as np
 import pandas as pd
 
 class Data:
-    def __init__(self, iterations: int, particle_amount: int) -> None:
+    def __init__(self) -> None:
         self.__particle_history: list[pd.DataFrame] = []
         self.__gbest_history: list[list[int]] = []
         self.__number_of_optimizations: int = 0
@@ -18,7 +18,11 @@ class Data:
         self.__number_of_optimizations += 1
 
     def create_spreadsheet(self) -> None:
-        pass
+        writer = pd.ExcelWriter('optimization_results.xlsx', engine='xlsxwriter')
+        
+        for i in range(self.__number_of_optimizations):
+            self.__particle_history[i].to_excel(excel_writer=writer, sheet_name=f'Optimization {i}')
+        writer.close()
 
     def print_optimization(self, optimization_index: int) -> None:
         pd.set_option('display.max_columns', None)
